created folder -atoms- where we can work on different components. moved all the pictures inside a new folder inside styles
