+++ src/components/atoms/logo/index.js
