--- src/components/tesstt/Tesstt.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Tesstt(props) {
-    return (
-        <div>
-            lalala
-        </div>
-    );
-}
-
-export default Tesstt;
